fix(authz): stabilize message deletion checks
