+++ cli/internal/core/authorization_consistency_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"hmans.de/chatto/internal/evtstream"
+	evtv1 "hmans.de/chatto/internal/pb/chatto/core/evt/v1"
 )
 
 func TestAuthorizeAtStableInputsRepeatsChangedDecision(t *testing.T) {
@@ -78,5 +79,84 @@ func TestAuthorityChangesDoNotWriteLegacyFence(t *testing.T) {
 	}
 	if after != before {
 		t.Fatalf("legacy authorization fence advanced from %d to %d", before, after)
+	}
+}
+
+func TestAuthorizedGroupMutationRechecksAfterPermissionRevocation(t *testing.T) {
+	core, _ := setupTestCore(t)
+	ctx := testContext(t)
+	actor, err := core.CreateUser(ctx, SystemActorID, "stable-group-manager", "Stable Group Manager", "password123")
+	if err != nil {
+		t.Fatalf("CreateUser: %v", err)
+	}
+	if err := core.GrantUserPermission(ctx, SystemActorID, actor.Id, PermRoomManage); err != nil {
+		t.Fatalf("GrantUserPermission room.manage: %v", err)
+	}
+	groups, err := core.ListRoomGroupsOrdered(ctx, KindChannel)
+	if err != nil || len(groups) == 0 {
+		t.Fatalf("ListRoomGroupsOrdered groups=%d err=%v", len(groups), err)
+	}
+	group := groups[0]
+	event := newEvent(actor.Id, &evtv1.Event{Event: &evtv1.Event_RoomGroupUpdated{
+		RoomGroupUpdated: &evtv1.RoomGroupUpdatedEvent{
+			GroupId:     group.GetId(),
+			Name:        "must-not-commit",
+			Description: group.GetDescription(),
+		},
+	}})
+
+	checks := 0
+	authorize := func() error {
+		checks++
+		if checks == 1 {
+			if err := core.ClearUserPermissionState(ctx, SystemActorID, actor.Id, PermRoomManage); err != nil {
+				return err
+			}
+			return nil
+		}
+		return core.requireCanManageRoomGroup(ctx, actor.Id, group.GetId())
+	}
+
+	if _, err := core.appendGroupLayoutMutation(ctx, evtstream.GroupAggregate(group.GetId()), event, authorize); !errors.Is(err, ErrPermissionDenied) {
+		t.Fatalf("appendGroupLayoutMutation error = %v, want ErrPermissionDenied", err)
+	}
+	updated, err := core.GetRoomGroup(ctx, group.GetId())
+	if err != nil {
+		t.Fatalf("GetRoomGroup: %v", err)
+	}
+	if updated.GetName() == "must-not-commit" {
+		t.Fatal("group mutation committed after room.manage was revoked")
+	}
+}
+
+func TestScopedPermissionMutationRechecksAfterRoleManageRevocation(t *testing.T) {
+	core, _ := setupTestCore(t)
+	ctx := testContext(t)
+	actor, err := core.CreateUser(ctx, SystemActorID, "stable-role-manager", "Stable Role Manager", "password123")
+	if err != nil {
+		t.Fatalf("CreateUser: %v", err)
+	}
+	if err := core.GrantUserPermission(ctx, SystemActorID, actor.Id, PermRoleManage); err != nil {
+		t.Fatalf("GrantUserPermission role.manage: %v", err)
+	}
+
+	checks := 0
+	authorize := func() error {
+		checks++
+		if checks == 1 {
+			if err := core.ClearUserPermissionState(ctx, SystemActorID, actor.Id, PermRoleManage); err != nil {
+				return err
+			}
+			return nil
+		}
+		return core.requireCanManageAdminRoles(ctx, actor.Id)
+	}
+
+	err = core.applyRolePermissionState(ctx, actor.Id, ScopeServer, "", RoleModerator, PermRoomCreate, PermissionStateAllow, authorize)
+	if !errors.Is(err, ErrPermissionDenied) {
+		t.Fatalf("applyRolePermissionState error = %v, want ErrPermissionDenied", err)
+	}
+	if got := core.rbacModel.decision(ScopeServer, "", RoleModerator, PermRoomCreate); got == DecisionAllow {
+		t.Fatal("permission mutation committed after role.manage was revoked")
 	}
 }
